refactor(agent-adapter): split codex's local JSONL history parsing into codex-history.ts
The JSONL transcript scanning/parsing free functions (readCodexIndex,
collectJsonlFiles, readJsonlFile, readCodexTranscriptSummary,
mapCodexHistoryEvents, etc.) share no state with the live CodexAdapter class
and were about half of codex.ts by size. Move them as-is into
native/codex-history.ts; codex.ts now imports what listHistory/readHistory need.
