+++ packages/agent-adapter/src/native/codex.ts
@@ -1,18 +1,10 @@
-import type { Stats } from 'node:fs';
-import { readdir, readFile, stat } from 'node:fs/promises';
-import { homedir } from 'node:os';
-import { basename, join } from 'node:path';
-import { env } from 'node:process';
 import type {
   AgentHistoryCapabilities,
-  AgentHistoryEvent,
-  AgentHistoryId,
   AgentHistoryListOptions,
   AgentHistoryListResult,
   AgentHistoryReadOptions,
   AgentHistoryReadResult,
   AgentHistoryResumeOptions,
-  AgentHistorySession,
   ContentBlock,
   StartOptions,
   TokenUsage,
@@ -20,30 +12,24 @@ import type {
   ToolCallStatus,
 } from '@linkcode/schema';
 import type { ThreadEvent, ThreadItem, ThreadOptions, Usage } from '@openai/codex-sdk';
-import { appendArrayInPlace } from 'foxts/append-array-in-place';
 import { extractErrorMessage } from 'foxts/extract-error-message';
-import { invariant, not } from 'foxts/guard';
+import { invariant } from 'foxts/guard';
 import { BaseAgentAdapter } from '../base';
-import {
-  asHistoryId,
-  boundedLimit,
-  compactRecord,
-  cursorFromTotal,
-  cursorOffset,
-  firstText,
-  isRecord,
-  recordField,
-  stringField,
-  textFromUnknown,
-  textHistoryEvent,
-  timestampMs,
-} from '../history-util';
+import { asHistoryId, boundedLimit, cursorFromTotal, cursorOffset } from '../history-util';
 import { contentToText } from '../util';
+import {
+  codexIndexEntryToSession,
+  codexSummaryToSession,
+  findCodexTranscript,
+  mapCodexHistoryEvents,
+  readCodexIndex,
+  readCodexTranscriptSummaries,
+  readJsonlFile,
+} from './codex-history';
 
 type CodexModule = typeof import('@openai/codex-sdk');
 type CodexInstance = InstanceType<CodexModule['Codex']>;
 type CodexThread = ReturnType<CodexInstance['startThread']>;
-const WHITESPACE_RUN_RE = /\s+/g;
 
 /** Map a Codex command/MCP status to our ToolCallStatus. */
 export function mapCodexStatus(status: 'in_progress' | 'completed' | 'failed'): ToolCallStatus {
@@ -280,265 +266,4 @@ export class CodexAdapter extends BaseAgentAdapter {
 function textContent(text: string): ToolCallContent[] {
   if (text.length === 0) return [];
   return [{ type: 'content', content: { type: 'text', text } }];
-}
-
-type JsonRecord = Record<string, unknown>;
-
-interface CodexIndexEntry {
-  id: string;
-  title?: string;
-  updatedAt?: number;
-}
-
-interface CodexTranscriptSummary {
-  id: string;
-  path?: string;
-  title?: string;
-  cwd?: string;
-  model?: string;
-  createdAt?: number;
-  updatedAt?: number;
-  messageCount?: number;
-  metadata?: Record<string, unknown>;
-}
-
-interface DirectoryEntry {
-  name: string;
-  isDirectory(): boolean;
-  isFile(): boolean;
-}
-
-function codexHome(): string {
-  return env.CODEX_HOME ?? join(homedir(), '.codex');
-}
-
-async function readCodexIndex(): Promise<Map<string, CodexIndexEntry>> {
-  const rows = await readJsonlFile(join(codexHome(), 'session_index.jsonl'));
-  const index = new Map<string, CodexIndexEntry>();
-  for (const row of rows) {
-    const id = stringField(row, 'id');
-    if (!id) continue;
-    index.set(id, {
-      id,
-      title: firstText(stringField(row, 'thread_name'), stringField(row, 'title')),
-      updatedAt: timestampMs(row.updated_at) ?? timestampMs(row.updatedAt),
-    });
-  }
-  return index;
-}
-
-async function readCodexTranscriptSummaries(
-  index: Map<string, CodexIndexEntry>,
-): Promise<CodexTranscriptSummary[]> {
-  const roots = [join(codexHome(), 'sessions'), join(codexHome(), 'archived_sessions')];
-  const fileSets = await Promise.all(roots.map((root) => collectJsonlFiles(root)));
-  const files = fileSets.flat();
-  const summaries = await Promise.all(files.map((file) => readCodexTranscriptSummary(file, index)));
-  return summaries.filter(not(undefined));
-}
-
-async function findCodexTranscript(
-  historyId: AgentHistoryId,
-): Promise<CodexTranscriptSummary | undefined> {
-  const index = await readCodexIndex();
-  const summaries = await readCodexTranscriptSummaries(index);
-  const id = historyId;
-  return summaries.find((summary) => summary.id === id || summary.path === id);
-}
-
-async function collectJsonlFiles(root: string, depth = 8): Promise<string[]> {
-  if (depth < 0) return [];
-  let entries: DirectoryEntry[];
-  try {
-    entries = await readdir(root, { withFileTypes: true, encoding: 'utf8' });
-  } catch {
-    return [];
-  }
-  const files: string[] = [];
-  const pendingDirs: Array<Promise<string[]>> = [];
-  for (const entry of entries) {
-    const path = join(root, entry.name);
-    if (entry.isDirectory()) pendingDirs.push(collectJsonlFiles(path, depth - 1));
-    else if (entry.isFile() && entry.name.endsWith('.jsonl')) files.push(path);
-  }
-  for (const nestedFiles of await Promise.all(pendingDirs)) appendArrayInPlace(files, nestedFiles);
-  return files;
-}
-
-async function readJsonlFile(path: string): Promise<JsonRecord[]> {
-  let raw: string;
-  try {
-    raw = await readFile(path, 'utf8');
-  } catch {
-    return [];
-  }
-  const rows: JsonRecord[] = [];
-  for (const line of raw.split('\n')) {
-    if (line.trim().length === 0) continue;
-    try {
-      const parsed: unknown = JSON.parse(line);
-      if (isRecord(parsed)) rows.push(parsed);
-    } catch {
-      // Ignore corrupt partial lines; Codex may be writing the active transcript.
-    }
-  }
-  return rows;
-}
-
-async function readCodexTranscriptSummary(
-  path: string,
-  index: Map<string, CodexIndexEntry>,
-): Promise<CodexTranscriptSummary | undefined> {
-  const [rows, fileStat] = await Promise.all([readJsonlFile(path), statOrUndefined(path)]);
-  if (rows.length === 0) return undefined;
-
-  let id: string | undefined;
-  let title: string | undefined;
-  let cwd: string | undefined;
-  let model: string | undefined;
-  let createdAt: number | undefined;
-  let updatedAt: number | undefined;
-  let firstUserText: string | undefined;
-  let firstAssistantText: string | undefined;
-  let messageCount = 0;
-  let cliVersion: string | undefined;
-  let originator: string | undefined;
-  let threadSource: string | undefined;
-  let modelProvider: string | undefined;
-  let gitBranch: string | undefined;
-
-  for (const row of rows) {
-    const rowType = stringField(row, 'type');
-    const rowTs = timestampMs(row.timestamp);
-    if (rowTs !== undefined) {
-      createdAt ??= rowTs;
-      updatedAt = Math.max(updatedAt ?? 0, rowTs);
-    }
-
-    const payload = recordField(row, 'payload');
-    if (!payload) continue;
-
-    switch (rowType) {
-      case 'session_meta': {
-        id = stringField(payload, 'id') ?? id;
-        cwd = stringField(payload, 'cwd') ?? cwd;
-        model = stringField(payload, 'model') ?? model;
-        originator = stringField(payload, 'originator') ?? originator;
-        threadSource = stringField(payload, 'thread_source') ?? threadSource;
-        cliVersion = stringField(payload, 'cli_version') ?? cliVersion;
-        modelProvider = stringField(payload, 'model_provider') ?? modelProvider;
-        const git = recordField(payload, 'git');
-        if (git) gitBranch = stringField(git, 'branch') ?? gitBranch;
-        createdAt = timestampMs(payload.timestamp) ?? createdAt;
-
-        break;
-      }
-      case 'turn_context': {
-        cwd = stringField(payload, 'cwd') ?? cwd;
-        model = stringField(payload, 'model') ?? model;
-        title = stringField(payload, 'summary') ?? title;
-
-        break;
-      }
-      case 'response_item': {
-        const role = stringField(payload, 'role');
-        if (role !== 'user' && role !== 'assistant') continue;
-        const text = textFromUnknown(payload);
-        if (text.trim().length === 0) continue;
-        messageCount += 1;
-        if (role === 'user') firstUserText ??= previewText(text);
-        else firstAssistantText ??= previewText(text);
-
-        break;
-      }
-      default:
-        break;
-    }
-  }
-
-  id ??= idFromFilename(path);
-  const indexEntry = index.get(id);
-  return {
-    id,
-    path,
-    title: firstText(indexEntry?.title, title, firstUserText, firstAssistantText),
-    cwd,
-    model,
-    createdAt,
-    updatedAt:
-      indexEntry?.updatedAt ?? updatedAt ?? (fileStat ? Math.trunc(fileStat.mtimeMs) : undefined),
-    messageCount,
-    metadata: compactRecord({
-      source: 'codex-local-jsonl',
-      transcriptPath: path,
-      fileSize: fileStat?.size,
-      cliVersion,
-      originator,
-      threadSource,
-      modelProvider,
-      gitBranch,
-    }),
-  };
-}
-
-async function statOrUndefined(path: string): Promise<Stats | undefined> {
-  try {
-    return await stat(path);
-  } catch {
-    // Transcript files can disappear while Codex is rotating or archiving sessions.
-  }
-}
-
-function codexSummaryToSession(summary: CodexTranscriptSummary): AgentHistorySession {
-  return {
-    historyId: asHistoryId(summary.id),
-    kind: 'codex',
-    title: summary.title,
-    cwd: summary.cwd,
-    model: summary.model,
-    createdAt: summary.createdAt,
-    updatedAt: summary.updatedAt,
-    messageCount: summary.messageCount,
-    metadata: summary.metadata,
-  };
-}
-
-function codexIndexEntryToSession(entry: CodexIndexEntry): AgentHistorySession {
-  return {
-    historyId: asHistoryId(entry.id),
-    kind: 'codex',
-    title: entry.title,
-    updatedAt: entry.updatedAt,
-    metadata: {
-      source: 'codex-session-index',
-      missingTranscript: true,
-    },
-  };
-}
-
-function mapCodexHistoryEvents(historyId: AgentHistoryId, rows: JsonRecord[]): AgentHistoryEvent[] {
-  const events: AgentHistoryEvent[] = [];
-  rows.forEach((row, index) => {
-    if (stringField(row, 'type') !== 'response_item') return;
-    const payload = recordField(row, 'payload');
-    if (!payload) return;
-    const role = stringField(payload, 'role');
-    if (role !== 'user' && role !== 'assistant') return;
-    const itemId =
-      stringField(payload, 'id') ?? stringField(row, 'id') ?? `${role}-${index.toString(36)}`;
-    const event = textHistoryEvent(historyId, role, itemId, payload, timestampMs(row.timestamp));
-    if (event) events.push(event);
-  });
-  return events;
-}
-
-function idFromFilename(path: string): string {
-  const name = basename(path, '.jsonl');
-  return name.length > 0 ? name : path;
-}
-
-function previewText(text: string): string {
-  const flat = text.replaceAll(WHITESPACE_RUN_RE, ' ').trim();
-  if (flat.length <= 120) return flat;
-  return `${flat.slice(0, 117)}...`;
 }
